+++ packages/transaction-controller/src/helpers/IncomingTransactionHelper.ts
@@ -1,32 +1,12 @@
 import type { AccountsController } from '@metamask/accounts-controller';
-import type { BlockTracker } from '@metamask/network-controller';
 import type { Hex } from '@metamask/utils';
-import { Mutex } from 'async-mutex';
+// This package purposefully relies on Node's EventEmitter module.
+// eslint-disable-next-line import/no-nodejs-modules
 import EventEmitter from 'events';
 
-import {
-  createModuleLogger,
-  incomingTransactionsLogger as log,
-} from '../logger';
+import { incomingTransactionsLogger as log } from '../logger';
 import type { RemoteTransactionSource, TransactionMeta } from '../types';
 
-const RECENT_HISTORY_BLOCK_RANGE = 10;
-
-// TODO: Replace `any` with type
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const UPDATE_CHECKS: ((txMeta: TransactionMeta) => any)[] = [
-  (txMeta) => txMeta.status,
-  (txMeta) => txMeta.txParams.gasUsed,
-];
-
-/**
- * Configuration options for the IncomingTransactionHelper
- *
- * @property includeTokenTransfers - Whether or not to include ERC20 token transfers.
- * @property isEnabled - Whether or not incoming transaction retrieval is enabled.
- * @property queryEntireHistory - Whether to initially query the entire transaction history or only recent blocks.
- * @property updateTransactions - Whether to update local transactions using remote transaction data.
- */
 export type IncomingTransactionOptions = {
   includeTokenTransfers?: boolean;
   isEnabled?: () => boolean;
@@ -34,88 +14,80 @@ export type IncomingTransactionOptions = {
   updateTransactions?: boolean;
 };
 
+const INTERVAL = 1000 * 30; // 30 Seconds
+
 export class IncomingTransactionHelper {
   hub: EventEmitter;
 
-  #blockTracker: BlockTracker;
+  #getCache: () => Record<string, unknown>;
 
   #getCurrentAccount: () => ReturnType<
     AccountsController['getSelectedAccount']
   >;
 
-  #getLastFetchedBlockNumbers: () => Record<string, number>;
+  #getChainIds: () => Hex[];
 
   #getLocalTransactions: () => TransactionMeta[];
 
-  #getChainId: () => Hex;
+  #includeTokenTransfers?: boolean;
 
   #isEnabled: () => boolean;
 
   #isRunning: boolean;
 
-  #log: debug.Debugger;
-
-  #mutex = new Mutex();
-
-  #onLatestBlock: (blockNumberHex: Hex) => Promise<void>;
-
-  #queryEntireHistory: boolean;
+  #queryEntireHistory?: boolean;
 
   #remoteTransactionSource: RemoteTransactionSource;
 
-  #transactionLimit?: number;
+  #timeoutId?: unknown;
 
-  #updateTransactions: boolean;
+  #trimTransactions: (transactions: TransactionMeta[]) => TransactionMeta[];
+
+  #updateCache: (fn: (cache: Record<string, unknown>) => void) => void;
+
+  #updateTransactions?: boolean;
 
   constructor({
-    blockTracker,
+    getCache,
     getCurrentAccount,
-    getLastFetchedBlockNumbers,
+    getChainIds,
     getLocalTransactions,
-    getChainId,
+    includeTokenTransfers,
     isEnabled,
     queryEntireHistory,
     remoteTransactionSource,
-    transactionLimit,
+    trimTransactions,
+    updateCache,
     updateTransactions,
   }: {
-    blockTracker: BlockTracker;
+    getCache: () => Record<string, unknown>;
     getCurrentAccount: () => ReturnType<
       AccountsController['getSelectedAccount']
     >;
-    getLastFetchedBlockNumbers: () => Record<string, number>;
-    getLocalTransactions?: () => TransactionMeta[];
-    getChainId: () => Hex;
+    getChainIds: () => Hex[];
+    getLocalTransactions: () => TransactionMeta[];
+    includeTokenTransfers?: boolean;
     isEnabled?: () => boolean;
     queryEntireHistory?: boolean;
     remoteTransactionSource: RemoteTransactionSource;
-    transactionLimit?: number;
+    trimTransactions: (transactions: TransactionMeta[]) => TransactionMeta[];
+    updateCache: (fn: (cache: Record<string, unknown>) => void) => void;
     updateTransactions?: boolean;
   }) {
     this.hub = new EventEmitter();
 
-    this.#blockTracker = blockTracker;
+    this.#getCache = getCache;
     this.#getCurrentAccount = getCurrentAccount;
-    this.#getLastFetchedBlockNumbers = getLastFetchedBlockNumbers;
-    this.#getLocalTransactions = getLocalTransactions || (() => []);
-    this.#getChainId = getChainId;
+    this.#getChainIds = getChainIds;
+    this.#getLocalTransactions = getLocalTransactions;
+    this.#includeTokenTransfers = includeTokenTransfers;
     this.#isEnabled = isEnabled ?? (() => true);
     this.#isRunning = false;
-    this.#log = createModuleLogger(log, getChainId());
-    this.#queryEntireHistory = queryEntireHistory ?? true;
+    this.#queryEntireHistory = queryEntireHistory;
     this.#remoteTransactionSource = remoteTransactionSource;
-    this.#transactionLimit = transactionLimit;
-    this.#updateTransactions = updateTransactions ?? false;
-
-    // Using a property instead of a method to provide a listener reference
-    // with the correct scope that we can remove later if stopped.
-    this.#onLatestBlock = async (blockNumberHex: Hex) => {
-      try {
-        await this.update(blockNumberHex);
-      } catch (error) {
-        console.error('Error while checking incoming transactions', error);
-      }
-    };
+    this.#trimTransactions = trimTransactions;
+    this.#updateCache = updateCache;
+    this.#updateTransactions = updateTransactions;
   }
 
   start() {
@@ -127,217 +99,147 @@ export class IncomingTransactionHelper {
       return;
     }
 
-    this.#log('Starting polling');
+    log('Starting polling');
 
-    // TODO: Either fix this lint violation or explain why it's necessary to ignore.
     // eslint-disable-next-line @typescript-eslint/no-misused-promises
-    this.#blockTracker.addListener('latest', this.#onLatestBlock);
+    this.#timeoutId = setTimeout(() => this.#onInterval(), INTERVAL);
     this.#isRunning = true;
+
+    log('Started polling');
   }
 
   stop() {
+    if (this.#timeoutId) {
+      clearTimeout(this.#timeoutId as number);
+    }
+
     if (!this.#isRunning) {
       return;
     }
 
-    // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-    // eslint-disable-next-line @typescript-eslint/no-misused-promises
-    this.#blockTracker.removeListener('latest', this.#onLatestBlock);
     this.#isRunning = false;
 
-    this.#log('Stopped polling');
+    log('Stopped polling');
   }
 
-  async update(latestBlockNumberHex?: Hex): Promise<void> {
-    const releaseLock = await this.#mutex.acquire();
+  async #onInterval() {
+    try {
+      await this.update({ isInterval: true });
+    } catch (error) {
+      console.error('Error while checking incoming transactions', error);
+    }
 
-    this.#log('Checking');
+    if (this.#isRunning) {
+      // eslint-disable-next-line @typescript-eslint/no-misused-promises
+      this.#timeoutId = setTimeout(() => this.#onInterval(), INTERVAL);
+    }
+  }
+
+  async update({ isInterval }: { isInterval?: boolean } = {}): Promise<void> {
+    log('Checking for incoming transactions', {
+      isInterval: Boolean(isInterval),
+    });
+
+    if (!this.#canStart()) {
+      return;
+    }
+
+    const account = this.#getCurrentAccount();
+    const chainIds = this.#getChainIds();
+    const cache = this.#getCache();
+    const includeTokenTransfers = this.#includeTokenTransfers ?? true;
+    const queryEntireHistory = this.#queryEntireHistory ?? true;
+    const updateTransactions = this.#updateTransactions ?? false;
+
+    let remoteTransactions: TransactionMeta[] = [];
 
     try {
-      if (!this.#canStart()) {
-        return;
-      }
-
-      const latestBlockNumber = parseInt(
-        latestBlockNumberHex || (await this.#blockTracker.getLatestBlock()),
-        16,
-      );
-
-      const additionalLastFetchedKeys =
-        this.#remoteTransactionSource.getLastBlockVariations?.() ?? [];
-
-      const fromBlock = this.#getFromBlock(latestBlockNumber);
-      const account = this.#getCurrentAccount();
-      const chainId = this.#getChainId();
-
-      let remoteTransactions = [];
-
-      try {
-        remoteTransactions =
-          await this.#remoteTransactionSource.fetchTransactions({
-            address: account.address,
-            chainId,
-            fromBlock,
-            limit: this.#transactionLimit,
-          });
-        // TODO: Replace `any` with type
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      } catch (error: any) {
-        this.#log('Error while fetching remote transactions', error);
-        return;
-      }
-
-      if (!this.#updateTransactions) {
-        const address = account.address.toLowerCase();
-        remoteTransactions = remoteTransactions.filter(
-          (tx) => tx.txParams.to?.toLowerCase() === address,
-        );
-      }
-
-      const localTransactions = !this.#updateTransactions
-        ? []
-        : this.#getLocalTransactions();
-
-      const newTransactions = this.#getNewTransactions(
-        remoteTransactions,
-        localTransactions,
-      );
-
-      const updatedTransactions = this.#getUpdatedTransactions(
-        remoteTransactions,
-        localTransactions,
-      );
-
-      if (newTransactions.length > 0 || updatedTransactions.length > 0) {
-        this.#sortTransactionsByTime(newTransactions);
-        this.#sortTransactionsByTime(updatedTransactions);
-
-        this.#log('Found incoming transactions', {
-          new: newTransactions,
-          updated: updatedTransactions,
+      remoteTransactions =
+        await this.#remoteTransactionSource.fetchTransactions({
+          address: account.address as Hex,
+          cache,
+          chainIds,
+          includeTokenTransfers,
+          queryEntireHistory,
+          updateCache: this.#updateCache,
+          updateTransactions,
         });
-
-        this.hub.emit('transactions', {
-          added: newTransactions,
-          updated: updatedTransactions,
-        });
-      }
-      this.#updateLastFetchedBlockNumber(
-        remoteTransactions,
-        additionalLastFetchedKeys,
-      );
-    } finally {
-      releaseLock();
+    } catch (error: unknown) {
+      log('Error while fetching remote transactions', error);
+      return;
     }
+
+    if (!remoteTransactions.length) {
+      return;
+    }
+
+    this.#sortTransactionsByTime(remoteTransactions);
+
+    log(
+      'Found potential transactions',
+      remoteTransactions.length,
+      remoteTransactions,
+    );
+
+    const localTransactions = this.#getLocalTransactions();
+
+    const uniqueTransactions = remoteTransactions.filter(
+      (tx) =>
+        !localTransactions.some(
+          (currentTx) =>
+            currentTx.hash?.toLowerCase() === tx.hash?.toLowerCase() &&
+            currentTx.txParams.from?.toLowerCase() ===
+              tx.txParams.from?.toLowerCase(),
+        ),
+    );
+
+    if (!uniqueTransactions.length) {
+      log('All transactions are already known');
+      return;
+    }
+
+    log(
+      'Found unique transactions',
+      uniqueTransactions.length,
+      uniqueTransactions,
+    );
+
+    const trimmedTransactions = this.#trimTransactions([
+      ...uniqueTransactions,
+      ...localTransactions,
+    ]);
+
+    const uniqueTransactionIds = uniqueTransactions.map((tx) => tx.id);
+
+    const newTransactions = trimmedTransactions.filter((tx) =>
+      uniqueTransactionIds.includes(tx.id),
+    );
+
+    if (!newTransactions.length) {
+      log('All unique transactions truncated due to limit');
+      return;
+    }
+
+    log('Adding new transactions', newTransactions.length, newTransactions);
+
+    this.hub.emit('transactions', newTransactions);
   }
 
   #sortTransactionsByTime(transactions: TransactionMeta[]) {
     transactions.sort((a, b) => (a.time < b.time ? -1 : 1));
   }
 
-  #getNewTransactions(
-    remoteTxs: TransactionMeta[],
-    localTxs: TransactionMeta[],
-  ): TransactionMeta[] {
-    return remoteTxs.filter(
-      (tx) => !localTxs.some(({ hash }) => hash === tx.hash),
-    );
-  }
-
-  #getUpdatedTransactions(
-    remoteTxs: TransactionMeta[],
-    localTxs: TransactionMeta[],
-  ): TransactionMeta[] {
-    return remoteTxs.filter((remoteTx) =>
-      localTxs.some(
-        (localTx) =>
-          remoteTx.hash === localTx.hash &&
-          this.#isTransactionOutdated(remoteTx, localTx),
-      ),
-    );
-  }
-
-  #isTransactionOutdated(
-    remoteTx: TransactionMeta,
-    localTx: TransactionMeta,
-  ): boolean {
-    return UPDATE_CHECKS.some(
-      (getValue) => getValue(remoteTx) !== getValue(localTx),
-    );
-  }
-
-  #getLastFetchedBlockNumberDec(): number {
-    const additionalLastFetchedKeys =
-      this.#remoteTransactionSource.getLastBlockVariations?.() ?? [];
-    const lastFetchedKey = this.#getBlockNumberKey(additionalLastFetchedKeys);
-    const lastFetchedBlockNumbers = this.#getLastFetchedBlockNumbers();
-    return lastFetchedBlockNumbers[lastFetchedKey];
-  }
-
-  #getFromBlock(latestBlockNumber: number): number | undefined {
-    const lastFetchedBlockNumber = this.#getLastFetchedBlockNumberDec();
-
-    if (lastFetchedBlockNumber) {
-      return lastFetchedBlockNumber + 1;
-    }
-
-    return this.#queryEntireHistory
-      ? undefined
-      : latestBlockNumber - RECENT_HISTORY_BLOCK_RANGE;
-  }
-
-  #updateLastFetchedBlockNumber(
-    remoteTxs: TransactionMeta[],
-    additionalKeys: string[],
-  ) {
-    let lastFetchedBlockNumber = -1;
-
-    for (const tx of remoteTxs) {
-      const currentBlockNumberValue = tx.blockNumber
-        ? parseInt(tx.blockNumber, 10)
-        : -1;
-
-      lastFetchedBlockNumber = Math.max(
-        lastFetchedBlockNumber,
-        currentBlockNumberValue,
-      );
-    }
-
-    if (lastFetchedBlockNumber === -1) {
-      return;
-    }
-
-    const lastFetchedKey = this.#getBlockNumberKey(additionalKeys);
-    const lastFetchedBlockNumbers = this.#getLastFetchedBlockNumbers();
-    const previousValue = lastFetchedBlockNumbers[lastFetchedKey];
-
-    if (previousValue >= lastFetchedBlockNumber) {
-      return;
-    }
-
-    this.hub.emit('updatedLastFetchedBlockNumbers', {
-      lastFetchedBlockNumbers: {
-        ...lastFetchedBlockNumbers,
-        [lastFetchedKey]: lastFetchedBlockNumber,
-      },
-      blockNumber: lastFetchedBlockNumber,
-    });
-  }
-
-  #getBlockNumberKey(additionalKeys: string[]): string {
-    const chainId = this.#getChainId();
-    const currentAccount = this.#getCurrentAccount()?.address.toLowerCase();
-
-    return [chainId, currentAccount, ...additionalKeys].join('#');
-  }
-
   #canStart(): boolean {
     const isEnabled = this.#isEnabled();
-    const chainId = this.#getChainId();
+    const chainIds = this.#getChainIds();
 
-    const isSupportedNetwork =
-      this.#remoteTransactionSource.isSupportedNetwork(chainId);
+    const supportedChainIds =
+      this.#remoteTransactionSource.getSupportedChains();
 
-    return isEnabled && isSupportedNetwork;
+    const isAnyChainSupported = chainIds.some((chainId) =>
+      supportedChainIds.includes(chainId),
+    );
+
+    return isEnabled && isAnyChainSupported;
   }
 }